session_catalog: carve replay.rs (pure move)
Session-log replay leaves mod.rs: JSONL-to-entry conversion, replay
metadata/ids, context-snapshot replay entries, detail paging, websocket
bootstrap prep — with its tests, verbatim. SessionDetailPageEntries
fields widened (accessed from mod.rs). Purity-checked; session_catalog
tests conserved at 104.
